+++ kendotournament-core/src/main/java/com/softwaremagico/ktg/database/FightPool.java
@@ -31,7 +31,9 @@ public class FightPool extends TournamentDependentPool<Fight> {
 
     @Override
     protected HashMap<String, Fight> getFromDatabase(Tournament tournament) {
+        DatabaseConnection.getInstance().connect();
         List<Fight> fights = DatabaseConnection.getConnection().getDatabase().getFights(tournament);
+        DatabaseConnection.getInstance().disconnect();
         HashMap<String, Fight> hashMap = new HashMap<>();
         for (Fight fight : fights) {
             hashMap.put(getId(fight), fight);
@@ -41,17 +43,23 @@ public class FightPool extends TournamentDependentPool<Fight> {
 
     @Override
     protected void storeInDatabase(Tournament tournament, List<Fight> elementsToStore) {
+        if (elementsToStore.size() > 0) {
         DatabaseConnection.getConnection().getDatabase().addFights(elementsToStore);
+        }
     }
 
     @Override
     protected void removeFromDatabase(Tournament tournament, List<Fight> elementsToDelete) {
+         if (elementsToDelete.size() > 0) {
         DatabaseConnection.getConnection().getDatabase().removeFights(elementsToDelete);
+         }
     }
 
     @Override
     protected void updateDatabase(Tournament tournament, HashMap<Fight, Fight> elementsToUpdate) {
+         if (elementsToUpdate.size() > 0) {
         DatabaseConnection.getConnection().getDatabase().updateFights(elementsToUpdate);
+         }
     }
 
     @Override
